+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginSkeleton/AbstractMainPlugin.php
@@ -428,8 +428,9 @@ abstract class AbstractMainPlugin extends AbstractPlugin implements MainPluginIn
                  * @see layers/GatoGraphQLForWP/plugins/gatographql/src/Marketplace/LicenseValidationService.php `activateDeactivateValidateGatoGraphQLCommercialExtensions`
                  */
                 $userSettingsManager = UserSettingsManagerFacade::getInstance();
-                if ($userSettingsManager->getLicenseActivationTimestamp() !== null) {
-                    $userSettingsManager->removeLicenseActivationTimestamp();
+                $justActivatedLicenseExtensionNames = $userSettingsManager->getJustActivatedLicenseTransientExtensionNames();
+                if ($justActivatedLicenseExtensionNames !== null && $justActivatedLicenseExtensionNames !== []) {
+                    $userSettingsManager->removeJustActivatedLicenseTransient();
                     $this->enqueueFlushRewriteRules();
 
                     /**
@@ -438,7 +439,7 @@ abstract class AbstractMainPlugin extends AbstractPlugin implements MainPluginIn
                      */
                     add_action(
                         PluginAppHooks::INITIALIZE_APP,
-                        function (string $pluginAppGraphQLServerName) use ($registeredExtensionBaseNameInstances): void {
+                        function (string $pluginAppGraphQLServerName) use ($registeredExtensionBaseNameInstances, $justActivatedLicenseExtensionNames): void {
                             if (
                                 $pluginAppGraphQLServerName === PluginAppGraphQLServerNames::INTERNAL
                                 || $this->initializationException !== null
@@ -446,9 +447,14 @@ abstract class AbstractMainPlugin extends AbstractPlugin implements MainPluginIn
                                 return;
                             }
 
-                            $this->anyCommercialLicenseJustActivated();
+                            if (in_array($this->pluginSlug, $justActivatedLicenseExtensionNames)) {
+                                $this->isLicenseJustActivated();
+                            }
                             foreach ($registeredExtensionBaseNameInstances as $extensionInstance) {
-                                $extensionInstance->anyCommercialLicenseJustActivated();
+                                if (!in_array($extensionInstance->getPluginSlug(), $justActivatedLicenseExtensionNames)) {
+                                    continue;
+                                }
+                                $extensionInstance->isLicenseJustActivated();
                             }
                         },
                         PluginLifecyclePriorities::AFTER_EVERYTHING