+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginSkeleton/AbstractPlugin.php
@@ -452,7 +452,7 @@ abstract class AbstractPlugin implements PluginInterface
      * Allow to install plugin setup data after a
      * commercial license has been activated
      */
-    public function anyCommercialLicenseJustActivated(): void
+    public function isLicenseJustActivated(): void
     {
         // Override
     }