+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginSkeleton/PluginInterface.php
@@ -30,7 +30,7 @@ interface PluginInterface
      * Allow to install plugin setup data after
      * a commercial license has been activated
      */
-    public function anyCommercialLicenseJustActivated(): void;
+    public function isLicenseJustActivated(): void;
 
     /**
      * Plugin name