+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Settings/UserSettingsManager.php
@@ -8,6 +8,7 @@ use GatoGraphQL\GatoGraphQL\Facades\Registries\SystemModuleRegistryFacade;
 use GatoGraphQL\GatoGraphQL\Facades\Registries\SystemSettingsCategoryRegistryFacade;
 use GatoGraphQL\GatoGraphQL\Facades\Settings\OptionNamespacerFacade;
 use GatoGraphQL\GatoGraphQL\Facades\TimestampSettingsManagerFacade;
+use GatoGraphQL\GatoGraphQL\Facades\TransientSettingsManagerFacade;
 
 use function get_option;
 use function uniqid;
@@ -17,9 +18,8 @@ class UserSettingsManager implements UserSettingsManagerInterface
 {
     private const TIMESTAMP_CONTAINER = 'container';
     private const TIMESTAMP_OPERATIONAL = 'operational';
+    private const TRANSIENT_LICENSE_ACTIVATION = 'license-activation';
     private const TIMESTAMP_LICENSE_CHECK = 'license-check';
-    private const TIMESTAMP_LICENSE_ACTIVATION = 'license-activation';
-
     /**
      * Cache the values in memory
      *
@@ -28,11 +28,16 @@ class UserSettingsManager implements UserSettingsManagerInterface
     protected array $options = [];
 
     private ?TimestampSettingsManagerInterface $timestampSettingsManager = null;
+    private ?TransientSettingsManagerInterface $transientSettingsManager = null;
     private ?OptionNamespacerInterface $optionNamespacer = null;
 
     final protected function getTimestampSettingsManager(): TimestampSettingsManagerInterface
     {
         return $this->timestampSettingsManager ??= TimestampSettingsManagerFacade::getInstance();
+    }
+    final protected function getTransientSettingsManager(): TransientSettingsManagerInterface
+    {
+        return $this->transientSettingsManager ??= TransientSettingsManagerFacade::getInstance();
     }
     final protected function getOptionNamespacer(): OptionNamespacerInterface
     {
@@ -169,30 +174,38 @@ class UserSettingsManager implements UserSettingsManagerInterface
     }
 
     /**
-     * Timestamp of the latest activation of any commercial license
+     * Retrieve the extension names whose commercial license has
+     * just been activated.
+     *
+     * @return string[]|null The license-just-activated extension names
      */
-    public function getLicenseActivationTimestamp(): ?int
+    public function getJustActivatedLicenseTransientExtensionNames(): ?array
     {
-        return $this->getTimestamp(self::TIMESTAMP_LICENSE_ACTIVATION);
+        return $this->getTransientSettingsManager()->getTransient(self::TRANSIENT_LICENSE_ACTIVATION);
     }
 
     /**
-     * Store the current time to indicate the latest activation
-     * of any commercial license
+     * Store the extension names whose commercial license has
+     * just been activated.
+     *
+     * @param string[] $extensionSlugs
      */
-    public function storeLicenseActivationTimestamp(): void
+    public function storeJustActivatedLicenseTransient(array $extensionSlugs): void
     {
-        $this->storeTimestamp(self::TIMESTAMP_LICENSE_ACTIVATION);
+        $this->getTransientSettingsManager()->storeTransient(
+            self::TRANSIENT_LICENSE_ACTIVATION,
+            $extensionSlugs
+        );
     }
 
     /**
-     * Remove the flag to indicate the latest activation
-     * of any commercial license
+     * Remove the flag to indicate the extension names whose commercial
+     * license has just been activated.
      */
-    public function removeLicenseActivationTimestamp(): void
+    public function removeJustActivatedLicenseTransient(): void
     {
-        $this->getTimestampSettingsManager()->removeTimestamps([
-            self::TIMESTAMP_LICENSE_ACTIVATION,
+        $this->getTransientSettingsManager()->removeTransients([
+            self::TRANSIENT_LICENSE_ACTIVATION,
         ]);
     }