+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to `gatographql` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 12.0.0 - DATE
+
+### Added
+
+- Documentation for new Elementor Premium Extension (#3052)
+
+### Breaking changes
+
+- Plugin constructor signature receives nullable `commitHash` param (#3056)
+- Store the extension names whose license has just been activated (instead of a timestamp) to install setup data (#3057)
+- Renamed `getCommentPostID` to `getCommentCustomPostID` (#3073)
+
 ## 11.3.1 - 19/04/2025
 
 ### Improvements